refactor: split engine.go into engine.go + engine_eval.go (#335)
Extract runSingleEval() and its helper functions (buildRerunCommand,
evaluateToolUsage, readReviewedFiles, writeReviewedFiles, countPassed)
into engine_eval.go. Keep Run(), NewEngine*(), orchestration helpers
(dryRun, pairwise, config expansion), and type definitions in engine.go.

engine.go: ~860 lines (orchestration + types)
engine_eval.go: ~790 lines (single eval execution + file helpers)
